modified hyperparameter tuning script to work without access to original grid-search config
